Move paper case express pilot enrolment to mid handler
