+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/handler/PaperCaseCreateContestedMidHandler.java
@@ -9,17 +9,25 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.model.EventType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseData;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseDetails;
+import uk.gov.hmcts.reform.finrem.caseorchestration.service.FeatureToggleService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.InternationalPostalService;
+import uk.gov.hmcts.reform.finrem.caseorchestration.service.express.ExpressCaseService;
 
 @Slf4j
 @Service
 public class PaperCaseCreateContestedMidHandler extends FinremCallbackHandler {
 
+    private final FeatureToggleService featureToggleService;
+    private final ExpressCaseService expressCaseService;
     private final InternationalPostalService postalService;
 
     public PaperCaseCreateContestedMidHandler(FinremCaseDetailsMapper finremCaseDetailsMapper,
+                                              FeatureToggleService featureToggleService,
+                                              ExpressCaseService expressCaseService,
                                               InternationalPostalService postalService) {
         super(finremCaseDetailsMapper);
+        this.featureToggleService = featureToggleService;
+        this.expressCaseService = expressCaseService;
         this.postalService = postalService;
     }
 
@@ -35,6 +43,10 @@ public class PaperCaseCreateContestedMidHandler extends FinremCallbackHandler {
                                                                               String userAuthorisation) {
         FinremCaseDetails caseDetails = callbackRequest.getCaseDetails();
         log.info("Invoking contested event {} mid event callback", EventType.NEW_PAPER_CASE);
+
+        if (featureToggleService.isExpressPilotEnabled()) {
+            expressCaseService.setExpressCaseEnrollmentStatus(caseDetails.getData());
+        }
 
         FinremCaseData caseData = caseDetails.getData();
         return GenericAboutToStartOrSubmitCallbackResponse.<FinremCaseData>builder()
