+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/handler/PaperCaseCreateContestedMidHandlerTest.java
@@ -13,12 +13,17 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.model.EventType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseData;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseDetails;
+import uk.gov.hmcts.reform.finrem.caseorchestration.service.FeatureToggleService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.InternationalPostalService;
+import uk.gov.hmcts.reform.finrem.caseorchestration.service.express.ExpressCaseService;
 
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+import static uk.gov.hmcts.reform.finrem.caseorchestration.TestConstants.AUTH_TOKEN;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.ccd.callback.CallbackType.MID_EVENT;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.ccd.callback.CallbackType.SUBMITTED;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.model.EventType.CLOSE;
@@ -27,13 +32,17 @@ import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType.CO
 import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType.CONTESTED;
 
 @ExtendWith(MockitoExtension.class)
-class PaperCaseCreateContestedMidHandlerTest {
-    public static final String AUTH_TOKEN = "tokien:)";
+class PaperCaseCreateContestedMidHandlerTest  {
+
     @InjectMocks
     private PaperCaseCreateContestedMidHandler handler;
 
     @Mock
-    private InternationalPostalService postalService;
+    FeatureToggleService featureToggleService;
+    @Mock
+    ExpressCaseService expressCaseService;
+    @Mock
+    InternationalPostalService postalService;
 
     @ParameterizedTest
     @MethodSource
@@ -64,5 +73,36 @@ class PaperCaseCreateContestedMidHandlerTest {
             .caseDetails(FinremCaseDetails.builder().id(123L).caseType(CONTESTED)
                 .data(FinremCaseData.builder().ccdCaseType(CONTESTED).build()).build())
             .build();
+    }
+
+    @Test
+    void testGivenExpressPilotEnabled_ThenExpressCaseServiceCalled() {
+        FinremCallbackRequest callbackRequest = buildFinremCallbackRequest();
+        FinremCaseData caseData = callbackRequest.getCaseDetails().getData();
+        when(featureToggleService.isExpressPilotEnabled()).thenReturn(true);
+
+        handler.handle(callbackRequest, AUTH_TOKEN);
+
+        verify(expressCaseService).setExpressCaseEnrollmentStatus(caseData);
+    }
+
+    @Test
+    void testGivenExpressPilotDisabled_ThenExpressCaseServiceIsNotCalled() {
+        FinremCallbackRequest callbackRequest = buildFinremCallbackRequest();
+
+        FinremCaseData caseData = callbackRequest.getCaseDetails().getData();
+        when(featureToggleService.isExpressPilotEnabled()).thenReturn(false);
+
+        handler.handle(callbackRequest, AUTH_TOKEN);
+
+        verify(expressCaseService, never()).setExpressCaseEnrollmentStatus(caseData);
+    }
+
+    private FinremCallbackRequest buildFinremCallbackRequest() {
+        return FinremCallbackRequest.builder().caseDetails(
+            FinremCaseDetails.builder().data(
+                FinremCaseData.builder().ccdCaseType(CONTESTED).build()
+            ).build()
+        ).build();
     }
 }
